requirements.txt
annotated-types==0.7.0
anyio==4.8.0
backports.tarfile==1.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
docutils==0.21.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.2
jsonpatch>=1.33
jsonpointer>=3.0.0
keyring==25.6.0
langchain-core>=0.3.49
langchain-openai>=0.3.11
langsmith>=0.3.19
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.20
openai>=1.69.0
orjson==3.10.16
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
PyYAML>=6.0.2
readme_renderer==44.0
regex>=2024.11.6
requests>=2.32.3
requests-toolbelt>=1.0.0
rfc3986==2.0.0
rich==13.9.4
sniffio==1.3.1
tenacity>=9.0.0
tiktoken>=0.9.0
tqdm==4.67.1
twine==6.1.0
typing_extensions>=4.12.2
urllib3==2.3.0
zipp==3.21.0
zstandard>=0.23.0
